# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MissionLedger API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /donations:
    get:
      operationId: getDonations
      tags: [donations]
      summary: Get all donations
      responses:
        "200":
          description: List of donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Create a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationRequest"
      responses:
        "201":
          description: Donation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
  /donations/{id}:
    get:
      operationId: getDonation
      tags: [donations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    put:
      operationId: updateDonation
      tags: [donations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationRequest"
      responses:
        "200":
          description: Donation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    delete:
      operationId: deleteDonation
      tags: [donations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: Get all expenses
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /funds:
    get:
      operationId: getFunds
      tags: [funds]
      summary: Get all funds
      responses:
        "200":
          description: List of funds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fund"
    post:
      operationId: createFund
      tags: [funds]
      summary: Create a fund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFundRequest"
      responses:
        "201":
          description: Fund created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fund"
  /funds/{id}:
    put:
      operationId: updateFund
      tags: [funds]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFundRequest"
      responses:
        "200":
          description: Fund updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fund"
    delete:
      operationId: deleteFund
      tags: [funds]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
  /bills:
    get:
      operationId: getBills
      tags: [bills]
      summary: Get all bills
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    put:
      operationId: updateBill
      tags: [bills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "200":
          description: Bill updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
    delete:
      operationId: deleteBill
      tags: [bills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /bills/{id}/payments:
    post:
      operationId: createBillPayment
      tags: [bills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillPaymentRequest"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillPayment"
...
components:
  schemas:
...
    Donation:
      type: object
      properties:
        id:
          type: string
        donorName:
          type: string
        donorEmail:
          type: string
        amount:
          type: number
        date:
          type: string
        type:
          type: string
        fundId:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        cashAccountId:
          type: string
          nullable: true
        fund:
          $ref: "#/components/schemas/Fund"
          nullable: true
        account:
          $ref: "#/components/schemas/Account"
          nullable: true
        journalEntryId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, donorName, amount, date, type, createdAt, updatedAt]
    CreateDonationRequest:
      type: object
      properties:
        donorName:
          type: string
        donorEmail:
          type: string
        amount:
          type: number
        date:
          type: string
        type:
          type: string
        fundId:
          type: string
        accountId:
          type: string
        cashAccountId:
          type: string
        notes:
          type: string
      required: [donorName, amount, date, type]
    Expense:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
        category:
          type: string
        fundId:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        cashAccountId:
          type: string
          nullable: true
        vendorId:
          type: string
          nullable: true
        vendor:
          $ref: "#/components/schemas/Vendor"
          nullable: true
        fund:
          $ref: "#/components/schemas/Fund"
          nullable: true
        journalEntryId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, description, amount, date, category, createdAt, updatedAt]
    CreateExpenseRequest:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        date:
          type: string
        category:
          type: string
        fundId:
          type: string
        accountId:
          type: string
        cashAccountId:
          type: string
        vendorId:
          type: string
        notes:
          type: string
      required: [description, amount, date, category]
    Fund:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        balance:
          type: number
        totalDonations:
          type: number
        totalExpenses:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, isActive, createdAt, updatedAt]
    CreateFundRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
      required: [name]
    Account:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        parentId:
          type: string
          nullable: true
        balance:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, code, name, type, isActive, createdAt, updatedAt]
...
    Vendor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        taxId:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, isActive, createdAt, updatedAt]
...
    Bill:
      type: object
      properties:
        id:
          type: string
        vendorId:
          type: string
          nullable: true
        vendor:
          $ref: "#/components/schemas/Vendor"
          nullable: true
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
        status:
          type: string
        accountId:
          type: string
          nullable: true
        fundId:
          type: string
          nullable: true
        payments:
          type: array
          items:
            $ref: "#/components/schemas/BillPayment"
        paidAmount:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, description, amount, dueDate, status, createdAt, updatedAt]
    CreateBillRequest:
      type: object
      properties:
        vendorId:
          type: string
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
        accountId:
          type: string
        fundId:
          type: string
      required: [description, amount, dueDate]
    BillPayment:
      type: object
      properties:
        id:
          type: string
        billId:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, billId, amount, date, createdAt]
    CreateBillPaymentRequest:
      type: object
      properties:
        amount:
          type: number
        date:
          type: string
        cashAccountId:
          type: string
        notes:
          type: string
      required: [amount, date]
...